requirements.txt
supabase==2.13.0
python-dotenv==1.0.1
pandas==2.2.3
streamlit==1.43.2
altair==5.5.0
plotly==6.0.1
matplotlib==3.9.4


